moving signer utils to it's own file
